pyproject.toml
[tool.poetry]
name = "onyx"
version = "0.1.0"
description = ""
authors = ["Ed"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
dbt-core = "^1.9.0"
dbt-postgres = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
